feat/create notification banner cookie backend
