+++ dataworkspace/dataworkspace/templates/datasets/index.html
@@ -17,7 +17,6 @@
 
 {% block header %}
   <div class="home-header">
-
     {{ block.super }}
     {% if banner %}
       {% include 'partials/notification_banner.html' %}
